[Chart Injection] Add Metadata Text on chart highlight (#5176)
Adds the provenance name, source name and some text about DC processing
on the Block description for the highlight chart.

This refactors some of the code to fetch the metadata from the show
Metadata modal to make it shareable.


http://localhost:8080/explore#sv=LifeExpectancy_Person&p=country%2FITA___country/FRA&client=ui_related_topic&chartType=TIMELINE_WITH_HIGHLIGHT&unit=Euro&imp=WorldDevelopmentIndicators
([screenshot](https://screenshot.googleplex.com/7DbYaBYFLvzoRvM))
