+++ static/js/tools/shared/metadata/citations.ts
@@ -37,7 +37,8 @@ export interface CitationPart {
  */
 export function buildCitationParts(
   statVars: NamedNode[],
-  metadataMap: Record<string, StatVarMetadata[]>
+  metadataMap: Record<string, StatVarMetadata[]>,
+  skipUrls?: boolean
 ): CitationPart[] {
   const seen = new Set<string>();
   const parts: CitationPart[] = [];
@@ -60,16 +61,18 @@ export function buildCitationParts(
         : metadata.provenanceName;
 
       const part: CitationPart = { label };
-      if (metadata.provenanceUrl) {
+      if (!skipUrls && metadata.provenanceUrl) {
         part.url = metadata.provenanceUrl.replace(/\/$/, "");
       }
       parts.push(part);
     });
   });
 
-  parts.push({
-    label: intl.formatMessage(metadataComponentMessages.MinorProcessing),
-  });
+  if (parts.length !== 0) {
+    parts.push({
+      label: intl.formatMessage(metadataComponentMessages.MinorProcessing),
+    });
+  }
 
   return parts;
 }
